Show data trending

+++ app/src/main/java/com/edu/movie/data/model/MovieItem.kt
@@ -4,7 +4,7 @@ data class MovieItem(
     val id: Int?,
     val title: String?,
     val imageUrl: String?,
-    val rate: Int?
+    val rate: Double?
 )
 
 object MovieEntry {
